requirements.txt
#	說明 
#使用jetson xavier nx,(arm64)、Ubuntu 20.04.6
#python 3.8、jetpack 5.X系列、YoloV 7、
#cuda 11.4.r11.4 、 PyTorch 1.12.0a0+2c916ef.nv22.3
#
#	重點
#在使用pip升級或安裝套件時,可能會升級到cuda或PyTorch的相關依賴性套件
#因為jetson xavier nx有專用的 PyTorch 和 CUDA
#如果升級到了，可以嘗試跑一次之前可執行的檔案，看是否有出現報錯，確定是PyTorch或CUDA引起的錯誤，那重灌系統是((最快的方法))
#
#	重灌系統最快的原因
#想一個一個解決相關依賴性套件，不可能，因為他是依賴性套件，你的升、降、刪除、重載、都有可能會動到超過一個套件的版本
#就會陷入一個死循環，A卡B,B又卡C跟D，動了C，C又卡A
absl-py==2.1.0
altgraph==0.17.4
asttokens==2.4.1
autocommand==2.2.2
backcall==0.2.0
backports.tarfile==1.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docutils==0.20.1
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gpio==1.0.0
grpcio==1.66.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.3.1
ipython==8.12.3
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.1
Jetson.GPIO==2.1.6
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tomli==2.0.1
torch @ file:///home/wu/Downloads/torch-1.12.0a0%2B2c916ef.nv22.3-cp38-cp38-linux_aarch64.whl#sha256=710df772a94add16c2f8202b7ade726f1b03385d68b57dde359fd8216915558f
torchaudio==0.12.0
torchvision==0.13.0a0+da3794e
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
